+++ AdColony/AdColonyAdapterConfiguration.m
@@ -48,7 +48,7 @@ typedef NS_ENUM(NSInteger, AdColonyAdapterErrorCode) {
 #pragma mark - MPAdapterConfiguration
 
 - (NSString *)adapterVersion {
-    return @"4.1.3.0";
+    return @"4.1.4.0";
 }
 
 - (NSString *)biddingToken {