+++ Mintegral/MintegralBannerCustomEvent.h
@@ -1,4 +1,6 @@
-#if __has_include(<MoPubSDKFramework/MoPub.h>)
+#if __has_include(<MoPub/MoPub.h>)
+    #import <MoPub/MoPub.h>
+#elif __has_include(<MoPubSDKFramework/MoPub.h>)
     #import <MoPubSDKFramework/MoPub.h>
 #else
     #import "MPBannerCustomEvent.h"