+++ Mintegral/CHANGELOG.md
@@ -1,4 +1,17 @@
 ## Changelog
+ * 6.1.1.0.1
+    * Import `<MoPub/MoPub.h>` for banner, interstitial, and rewarded video adapter classes. 
+
+ * 6.1.1.0.0
+    * This version of the adapters has been certified with Mintegral 6.1.1.0 and MoPub 5.11.0.
+
+ * 6.1.0.0.0
+    * This version of the adapters has been certified with Mintegral 6.1.0.0.
+
+ * 6.0.0.0.0
+    * This version of the adapters has been certified with Mintegral 6.0.0.0.
+    * Add a mute API for interstitial and rewarded video. Unless muted via `[MintegralAdapterConfiguration setMute:YES]`, creatives play unmuted by default. 
+
  * 5.9.0.0.0
     * This version of the adapters has been certified with Mintegral 5.9.0.0